+++ src/Widgets/ModelMusclePlotPanel.hpp
@@ -5,17 +5,23 @@
 #include <string_view>
 
 namespace OpenSim { class ComponentPath; }
+namespace osc { class EditorAPI; }
 namespace osc { class UndoableModelStatePair; }
 
 namespace osc
 {
     class ModelMusclePlotPanel final {
     public:
-        ModelMusclePlotPanel(std::shared_ptr<UndoableModelStatePair>, std::string_view panelName);
-        ModelMusclePlotPanel(std::shared_ptr<UndoableModelStatePair>,
-                             std::string_view panelName,
-                             OpenSim::ComponentPath const& coordPath,
-                             OpenSim::ComponentPath const& musclePath);
+        ModelMusclePlotPanel(
+            EditorAPI*,
+            std::shared_ptr<UndoableModelStatePair>,
+            std::string_view panelName);
+        ModelMusclePlotPanel(
+            EditorAPI*,
+            std::shared_ptr<UndoableModelStatePair>,
+            std::string_view panelName,
+            OpenSim::ComponentPath const& coordPath,
+            OpenSim::ComponentPath const& musclePath);
         ModelMusclePlotPanel(ModelMusclePlotPanel const&) = delete;
         ModelMusclePlotPanel(ModelMusclePlotPanel&&) noexcept;
         ModelMusclePlotPanel& operator=(ModelMusclePlotPanel const&) = delete;